separate scheduleFetch function
