+++ src/interfaces/Lesson.ts
@@ -1,6 +1,6 @@
 type LessonClassname = 'lecture' | 'sem' | 'lab' | 'holiday'
 
-/** Массив, содержаищй информацию о паре
+/** Массив, содержащий информацию о паре
  * @param 0 - Номер дня недели [1-6]
  * @param 1 - Название предмета
  * @param 2 - Номер пары
