+++ src/scripts/scheduleFetch.ts
@@ -1,5 +1,4 @@
 import vueToDom from '@/helpers/vueToDom'
-import { createApp } from 'vue'
 import ScheduleFetch from '@/components/schedule/scheduleFetch.vue'
 
 window.onload = () => {
